celery/requirements.txt
python-dotenv
easydict
celery
redis
requests
selenium
fake_useragent
sentry_sdk
beautifulsoup4
lxml
cloudscraper
pymongo
pyOpenSSL